HOTT-1238 Terraform changes to reflect caching
There's been a number of issues with caching where pages that shouldn't
be cached were getting hits upon testing. I've altered the distribution
slightly on development whilst we diagmose this and implemented a number
of other things we should have here like pre-commit config. I've also
included a shell.nix for myself but this can be removed if not accepted
by others.

Signed-off-by: Chris Mills <7100370+chriscoffee@users.noreply.github.com>